refactor: extract GraphEngine trait to proximadb-graph-engine-traits leaf (ORION move 6a)
First slice of the ORION crate-move sub-cascade: the GraphEngine contract
(the 21-ref blocker in orion/) and its value types move from root
src/graph/engines/mod.rs to a new foundation leaf, so a graph engine (ORION)
can depend on the contract without a cyclic dependency on the root crate.

- New foundation crate proximadb-graph-engine-traits: the GraphEngine trait +
  its default impls + GraphEngineStats + MemoryUsage + GraphExportFormat +
  PersistenceConfig. Deps are foundation-only (graph-model, kernel,
  compression-types, async-trait) — no upward layering.
- root src/graph/engines/mod.rs: the trait + 4 types replaced by a
  `pub use proximadb_graph_engine_traits::{...}` shim, so all existing callers
  (~5 implementors, ~10 `dyn GraphEngine` sites) resolve unchanged. Drops the
  unused to_metrics_sample helpers (MetricsSample is control-tier; no callers)
  and the now-unused CompressionAlgorithm/MetricsSample/HashMap/Path imports.

Behavior-identical (pure relocation + re-export shim). Verified: clippy --lib
--bins (CI gate) + cargo check --tests clean; rustfmt 1.95.0 canonical;
panic-policy, layering, workspace-boundary guards clean (new foundation leaf
has no upward deps); leaf compiles standalone.
