+++ crates/foundation/proximadb-graph-engine-traits/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "proximadb-graph-engine-traits"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Foundation graph-engine contract: the GraphEngine trait + its stats/memory/persistence/export value types, so graph engines (ORION) and their callers can name the contract without a cyclic root dependency"
+
+[lib]
+name = "proximadb_graph_engine_traits"
+path = "src/lib.rs"
+
+[dependencies]
+async-trait.workspace = true
+# Foundation-tier: Node/Edge/NodeId/EdgeId value types named by the trait.
+proximadb-graph-model.workspace = true
+# Foundation-tier: ProximaDBError (the trait's Result error type).
+proximadb-kernel.workspace = true
+# Foundation-tier: CompressionAlgorithm (PersistenceConfig.compression field).
+proximadb-compression-types.workspace = true
+
+[lints]
+workspace = true
