+++ src/graph/engines/mod.rs
@@ -42,11 +42,7 @@ pub mod orion;
 
 use proximadb_kernel::error::ProximaDBError;
 type Result<T> = std::result::Result<T, ProximaDBError>;
-use crate::core::serialization::CompressionAlgorithm;
 use crate::graph::{Edge, EdgeId, Node, NodeId};
-use crate::metrics::collectors::MetricsSample;
-use std::collections::HashMap;
-use std::path::Path;
 use std::sync::Arc;
 
 /// Embedding storage mode for graph engines
@@ -193,342 +189,9 @@ mod embedding_mode_tests {
 /// Backwards-compat alias for [`GraphEngineStats`].
 pub type EngineStats = GraphEngineStats;
 
-/// Engine performance statistics integrated with unified metrics framework
-#[derive(Debug, Clone, Default)]
-pub struct GraphEngineStats {
-    /// Total number of nodes
-    pub node_count: usize,
-    /// Total number of edges
-    pub edge_count: usize,
-    /// Average degree (edges per node)
-    pub avg_degree: f64,
-    /// Number of connected components
-    pub connected_components: usize,
-    /// Total operations performed
-    pub total_operations: u64,
-    /// Cache hit ratio (if applicable)
-    pub cache_hit_ratio: f64,
-    /// Index efficiency metric
-    pub index_efficiency: f64,
-    /// Time spent in operations (microseconds)
-    pub total_time_us: u64,
-}
-
-impl GraphEngineStats {
-    /// Convert to unified metrics sample for integration with metrics framework
-    pub fn to_metrics_sample(&self, engine_name: &str) -> MetricsSample {
-        let mut values = HashMap::new();
-        values.insert("node_count".to_string(), self.node_count as f64);
-        values.insert("edge_count".to_string(), self.edge_count as f64);
-        values.insert("avg_degree".to_string(), self.avg_degree);
-        values.insert(
-            "connected_components".to_string(),
-            self.connected_components as f64,
-        );
-        values.insert("total_operations".to_string(), self.total_operations as f64);
-        values.insert("cache_hit_ratio".to_string(), self.cache_hit_ratio);
-        values.insert("index_efficiency".to_string(), self.index_efficiency);
-        values.insert("total_time_us".to_string(), self.total_time_us as f64);
-
-        MetricsSample {
-            timestamp: std::time::Instant::now(),
-            collector: format!("graph_engine_{}", engine_name),
-            values,
-        }
-    }
-}
-
-/// Memory usage metrics integrated with unified metrics framework
-#[derive(Debug, Clone, Default)]
-pub struct MemoryUsage {
-    /// Memory used by nodes (bytes)
-    pub nodes_memory: usize,
-    /// Memory used by edges (bytes)
-    pub edges_memory: usize,
-    /// Memory used by indexes (bytes)
-    pub indexes_memory: usize,
-    /// Memory used by caches (bytes)
-    pub cache_memory: usize,
-    /// Total memory used (bytes)
-    pub total_memory: usize,
-    /// Peak memory usage (bytes)
-    pub peak_memory: usize,
-}
-
-impl MemoryUsage {
-    /// Convert to unified metrics sample for integration with metrics framework
-    pub fn to_metrics_sample(&self, engine_name: &str) -> MetricsSample {
-        let mut values = HashMap::new();
-        values.insert("nodes_memory_bytes".to_string(), self.nodes_memory as f64);
-        values.insert("edges_memory_bytes".to_string(), self.edges_memory as f64);
-        values.insert(
-            "indexes_memory_bytes".to_string(),
-            self.indexes_memory as f64,
-        );
-        values.insert("cache_memory_bytes".to_string(), self.cache_memory as f64);
-        values.insert("total_memory_bytes".to_string(), self.total_memory as f64);
-        values.insert("peak_memory_bytes".to_string(), self.peak_memory as f64);
-
-        MetricsSample {
-            timestamp: std::time::Instant::now(),
-            collector: format!("graph_memory_{}", engine_name),
-            values,
-        }
-    }
-}
-
-/// Graph engine trait for common operations across all engines
-#[async_trait::async_trait]
-pub trait GraphEngine: Send + Sync {
-    /// Insert a node (async for durability - waits for WAL write)
-    async fn insert_node(&self, node: Node) -> Result<Arc<Node>>;
-
-    /// Get a node by ID
-    fn get_node(&self, id: &NodeId) -> Result<Option<Arc<Node>>>;
-
-    /// Update a node (async for durability - waits for WAL write)
-    async fn update_node(&self, node: Node) -> Result<Arc<Node>>;
-
-    /// Delete a node (async for durability - waits for WAL write)
-    async fn delete_node(&self, id: &NodeId) -> Result<Option<Arc<Node>>>;
-
-    /// Insert an edge (async for durability - waits for WAL write)
-    async fn insert_edge(&self, edge: Edge) -> Result<Arc<Edge>>;
-
-    /// Get an edge by ID
-    fn get_edge(&self, id: &EdgeId) -> Result<Option<Arc<Edge>>>;
-
-    /// Update an edge (async for durability - waits for WAL write)
-    async fn update_edge(&self, edge: Edge) -> Result<Arc<Edge>>;
-
-    /// Delete an edge (async for durability - waits for WAL write)
-    async fn delete_edge(&self, id: &EdgeId) -> Result<Option<Arc<Edge>>>;
-
-    /// Get outgoing edges from a node
-    fn get_outgoing_edges(
-        &self,
-        node_id: &NodeId,
-        edge_type: Option<&str>,
-    ) -> Result<Vec<Arc<Edge>>>;
-
-    /// Get incoming edges to a node
-    fn get_incoming_edges(
-        &self,
-        node_id: &NodeId,
-        edge_type: Option<&str>,
-    ) -> Result<Vec<Arc<Edge>>>;
-
-    /// Get neighbors of a node
-    fn get_neighbors(&self, node_id: &NodeId, edge_type: Option<&str>) -> Result<Vec<Arc<Node>>>;
-
-    /// Get nodes by label
-    fn get_nodes_by_label(&self, label: &str) -> Result<Vec<Arc<Node>>>;
-
-    /// Get total node count
-    fn node_count(&self) -> Result<usize>;
-
-    /// Get total edge count
-    fn edge_count(&self) -> Result<usize>;
-
-    /// Get all nodes
-    fn get_all_nodes(&self) -> Result<Vec<Arc<Node>>>;
-
-    /// Get all edges. Default returns empty (mocks / engines that don't persist);
-    /// real engines override it. Used by canonical-store recovery re-population
-    /// (TD-066 Part 2) to re-drive every recovered edge into the canonical store.
-    fn get_all_edges(&self) -> Result<Vec<Arc<Edge>>> {
-        Ok(Vec::new())
-    }
-
-    // ===== Performance & Benchmarking Methods =====
-
-    /// Get engine performance statistics
-    fn get_engine_stats(&self) -> Result<GraphEngineStats> {
-        // Default implementation for backward compatibility
-        Ok(GraphEngineStats::default())
-    }
-
-    /// Get memory usage metrics
-    fn get_memory_usage(&self) -> Result<MemoryUsage> {
-        Ok(MemoryUsage::default())
-    }
-
-    // ===== Bulk Operations for Benchmarking =====
-
-    /// Bulk insert nodes (optimized for batch operations)
-    async fn bulk_insert_nodes(&self, nodes: Vec<Node>) -> Result<Vec<Arc<Node>>> {
-        // Default implementation delegates to single insert
-        let mut results = Vec::with_capacity(nodes.len());
-        for node in nodes {
-            results.push(self.insert_node(node).await?);
-        }
-        Ok(results)
-    }
-
-    /// Bulk insert edges (optimized for batch operations)
-    async fn bulk_insert_edges(&self, edges: Vec<Edge>) -> Result<Vec<Arc<Edge>>> {
-        // Default implementation delegates to single insert
-        let mut results = Vec::with_capacity(edges.len());
-        for edge in edges {
-            results.push(self.insert_edge(edge).await?);
-        }
-        Ok(results)
-    }
-
-    /// Bulk delete nodes
-    async fn bulk_delete_nodes(&self, node_ids: Vec<NodeId>) -> Result<Vec<Option<Arc<Node>>>> {
-        let mut results = Vec::with_capacity(node_ids.len());
-        for id in node_ids {
-            results.push(self.delete_node(&id).await?);
-        }
-        Ok(results)
-    }
-
-    /// Bulk delete edges
-    async fn bulk_delete_edges(&self, edge_ids: Vec<EdgeId>) -> Result<Vec<Option<Arc<Edge>>>> {
-        let mut results = Vec::with_capacity(edge_ids.len());
-        for id in edge_ids {
-            results.push(self.delete_edge(&id).await?);
-        }
-        Ok(results)
-    }
-
-    // ===== Engine Optimization Methods =====
-
-    /// Optimize internal storage structures
-    fn optimize_storage(&self) -> Result<()> {
-        // Default no-op implementation
-        Ok(())
-    }
-
-    /// Rebuild indexes for better query performance
-    fn rebuild_indexes(&self) -> Result<()> {
-        // Default no-op implementation
-        Ok(())
-    }
-
-    /// Compact storage to reduce memory footprint
-    fn compact_storage(&self) -> Result<()> {
-        // Default no-op implementation
-        Ok(())
-    }
-
-    /// Clear all data (useful for benchmarking)
-    async fn clear_all(&self) -> Result<()> {
-        // Default implementation: delete all nodes and edges
-        let all_nodes = self.get_all_nodes()?;
-        for node in all_nodes {
-            self.delete_node(&node.id).await?;
-        }
-        Ok(())
-    }
-
-    // ===== Persistence Methods (Critical for Production) =====
-
-    /// Save graph snapshot to persistent storage
-    /// Uses UnifiedCachingFilesystem for cloud-native storage support
-    async fn save_snapshot(&self, _path: &Path) -> Result<()> {
-        // Default implementation - engines should override for optimal performance
-        Err(ProximaDBError::NotImplemented(
-            "Graph persistence not implemented for this engine".to_string(),
-        ))
-    }
-
-    /// Load graph from persistent storage
-    /// Restores graph state from a previous snapshot
-    async fn load_snapshot(&self, _path: &Path) -> Result<()> {
-        Err(ProximaDBError::NotImplemented(
-            "Graph persistence not implemented for this engine".to_string(),
-        ))
-    }
-
-    /// Create incremental checkpoint (for WAL-style persistence)
-    async fn checkpoint(&self) -> Result<()> {
-        // Default no-op - engines can implement incremental checkpointing
-        Ok(())
-    }
-
-    /// Export graph to standard format (GraphML, GEXF, etc.)
-    async fn export(&self, _format: GraphExportFormat, _path: &Path) -> Result<()> {
-        Err(ProximaDBError::NotImplemented(
-            "Graph export not implemented for this engine".to_string(),
-        ))
-    }
-
-    /// Import graph from standard format
-    async fn import(&self, _format: GraphExportFormat, _path: &Path) -> Result<()> {
-        Err(ProximaDBError::NotImplemented(
-            "Graph import not implemented for this engine".to_string(),
-        ))
-    }
-
-    /// Get persistence configuration for this engine
-    fn get_persistence_config(&self) -> PersistenceConfig {
-        PersistenceConfig::default()
-    }
-
-    /// Check if the engine supports persistence
-    fn supports_persistence(&self) -> bool {
-        false // Default: no persistence support
-    }
-}
-
-/// Graph export/import formats for interoperability
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum GraphExportFormat {
-    /// GraphML - XML-based format for graph exchange
-    GraphML,
-    /// GEXF - Graph Exchange XML Format
-    GEXF,
-    /// JSON - Native JSON representation
-    Json,
-    /// Binary - ProximaDB binary format (most efficient)
-    ProximaBinary,
-    /// CSV - Node/Edge lists in CSV format
-    Csv,
-}
-
-/// Persistence configuration for graph engines
-#[derive(Debug, Clone)]
-pub struct PersistenceConfig {
-    /// Enable automatic snapshots
-    pub auto_snapshot: bool,
-    /// Snapshot interval in seconds
-    pub snapshot_interval_secs: u64,
-    /// Enable WAL (Write-Ahead Log) for crash recovery
-    pub enable_wal: bool,
-    /// WAL directory path
-    pub wal_path: Option<String>,
-    /// Compression algorithm for snapshots (using unified compression module)
-    pub compression: CompressionAlgorithm,
-    /// Compression level (algorithm-specific, typically 1-9 or 1-22 for Zstd)
-    pub compression_level: i32,
-    /// Use cloud storage for snapshots (S3, Azure, GCS)
-    pub use_cloud_storage: bool,
-    /// Cloud storage bucket/container
-    pub cloud_storage_path: Option<String>,
-    /// Enable incremental snapshots (only save changes since last snapshot)
-    pub incremental_snapshots: bool,
-    /// Maximum number of snapshots to retain
-    pub max_snapshots: usize,
-}
-
-impl Default for PersistenceConfig {
-    fn default() -> Self {
-        Self {
-            auto_snapshot: false,
-            snapshot_interval_secs: 3600, // 1 hour default
-            enable_wal: false,
-            wal_path: None,
-            compression: CompressionAlgorithm::Zstd, // Best general-purpose compression
-            compression_level: 3,                    // Balanced speed/ratio for Zstd
-            use_cloud_storage: false,
-            cloud_storage_path: None,
-            incremental_snapshots: true,
-            max_snapshots: 10,
-        }
-    }
-}
+pub use proximadb_graph_engine_traits::{
+    GraphEngine, GraphEngineStats, GraphExportFormat, MemoryUsage, PersistenceConfig,
+};
 
 /// Engine type enumeration for factory creation
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
